requirements.txt
bottle  # for running web
gunicorn  #  webserver
requests  # for fetching data from other sub-micro-services
matplotlib
pylab  # needed?
mpl_toolkits.basemap  # how to import Basemap?
